+++ lib/controllers/in_app_review_tools_controller.dart
@@ -3,12 +3,12 @@ import 'package:brain_dev_tools/tools/constant.dart';
 import 'package:brain_dev_tools/tools/enum/enum.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:in_app_review/in_app_review.dart';
+// import 'package:in_app_review/in_app_review.dart';
 
 class InAppReviewToolsController  extends GetxController implements GetxService
 {
   //region [ ATTRIBUTS ]
-  final InAppReview inAppReview = InAppReview.instance;
+  // final InAppReview inAppReview = InAppReview.instance;
   String appStoreId = Constant.appStoreIdIdentifierCCG;
   String microsoftStoreId = '';
 
@@ -25,14 +25,14 @@ class InAppReviewToolsController  extends GetxController implements GetxService
     (<T>(T? o) => o!)(WidgetsBinding.instance).addPostFrameCallback((_) async
     {
       try {
-        final isAvailable = await inAppReview.isAvailable();
+        // final isAvailable = await inAppReview.isAvailable();
 
         // This plugin cannot be tested on Android by installing your app
         // locally. See https://github.com/britannio/in_app_review#testing for
         // more information.
-        _availability = isAvailable && !checkPlatform.isAndroid
-            ? Availability.available
-            : Availability.unavailable;
+        // _availability = isAvailable && !checkPlatform.isAndroid
+        //     ? Availability.available
+        //     : Availability.unavailable;
       } catch (_) {
         _availability = Availability.unavailable;
       }
@@ -42,12 +42,12 @@ class InAppReviewToolsController  extends GetxController implements GetxService
   //endregion
 
 //region [ ]
-  Future<void> requestReview() => inAppReview.requestReview();
-
-  Future<void> openStoreListing() => inAppReview.openStoreListing(
-    appStoreId: appStoreId,
-    microsoftStoreId: microsoftStoreId,
-  );
+//   Future<void> requestReview() => inAppReview.requestReview();
+//
+//   Future<void> openStoreListing() => inAppReview.openStoreListing(
+//     appStoreId: appStoreId,
+//     microsoftStoreId: microsoftStoreId,
+//   );
 //endregion
 
 //region [ ]